requirements.txt
matplotlib
jieba
wordcloud
numpy
pillow